app/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
astroid==2.4.2
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
elasticsearch==7.8.0
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.5.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
schedule==0.6.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
